# .loki/specs/openapi.yaml
openapi: 3.1.0
info:
  title: AgencyOS API
  description: Client Portal & Agency Fulfillment System API
  version: 1.0.0
  contact:
    name: AgencyOS
servers:
  - url: http://localhost:3000/api
    description: Development server
  - url: https://agencyos.example.com/api
    description: Production server
...
paths:
  /webhooks/woo:
    post:
      tags: [Webhooks]
      summary: WooCommerce subscription webhook
      description: |
        Handles subscription events from WooCommerce:
        - subscription.created: Creates new company and user
        - subscription.updated: Updates company status/plan
        - subscription.cancelled: Marks company as churned
      operationId: handleWooWebhook
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WooWebhookPayload'
      responses:
        '200':
          description: Webhook processed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookResponse'
        '400':
          description: Invalid payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Invalid webhook signature
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /requests:
    get:
      tags: [Requests]
      summary: List requests for a company
      description: Returns all requests for the authenticated user's company (or all if admin)
      operationId: listRequests
      parameters:
        - name: company_id
          in: query
          description: Filter by company (admin only)
          schema:
            type: string
            format: uuid
        - name: status
          in: query
          description: Filter by status
          schema:
            type: string
            enum: [queue, active, review, done]
      responses:
        '200':
          description: List of requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Request'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags: [Requests]
      summary: Create a new request
      description: Creates a new request in "queue" status
      operationId: createRequest
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRequestInput'
      responses:
        '201':
          description: Request created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Request'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Company is paused or churned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Supabase JWT token
...
  responses:
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
        code:
          type: string
        details:
          type: object
    WebhookResponse:
      type: object
      properties:
        success:
          type: boolean
        action:
          type: string
          enum: [created, updated, ignored]
        companyId:
          type: string
          format: uuid
    WooWebhookPayload:
      type: object
      required: [id, status, customer_id]
      properties:
        id:
          type: integer
          description: WooCommerce subscription ID
        status:
          type: string
          enum: [active, on-hold, cancelled, pending, expired]
        customer_id:
          type: integer
        billing:
          type: object
          properties:
            email:
              type: string
              format: email
            first_name:
              type: string
            last_name:
              type: string
            company:
              type: string
        line_items:
          type: array
          items:
            type: object
            properties:
              product_id:
                type: integer
              name:
                type: string
    Company:
      type: object
      required: [id, name, status, plan_tier, max_active_limit]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        status:
          type: string
          enum: [active, paused, churned]
        plan_tier:
          type: string
          enum: [standard, pro]
        max_active_limit:
          type: integer
          minimum: 1
        woo_customer_id:
          type: string
        stripe_customer_id:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        active_request_count:
          type: integer
          description: Current count of active requests
    Request:
      type: object
      required: [id, company_id, title, status, priority, created_at]
      properties:
        id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
          description: Markdown supported
        status:
          type: string
          enum: [queue, active, review, done]
        priority:
          type: string
          enum: [low, normal, high]
        assets_link:
          type: string
          format: uri
          description: Google Drive / Dropbox URL
        video_brief:
          type: string
          format: uri
          description: Loom URL
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        company:
          $ref: '#/components/schemas/Company'
    CreateRequestInput:
      type: object
      required: [title]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
        description:
          type: string
        priority:
          type: string
          enum: [low, normal, high]
          default: normal
        assets_link:
          type: string
          format: uri
        video_brief:
          type: string
          format: uri
...
security:
  - bearerAuth: []